trading-platform/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2

pydantic-settings==2.5.2


pytest==8.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==3.2.2


redis==5.0.1
hiredis==2.3.2

web3==6.20.2
eth-account==0.11.2
py-solc-x==2.0.3

